Search component moved to sub directory
