+++ components/Search/Search.vue
